+++ src/sprightly_inventory/lib/data/tables/app_tables.dart
@@ -1,0 +1,2 @@
+export 'custom_property.dart';
+export 'member.dart';
