+++ src/sprightly_inventory/lib/data/tables/setup_tables.dart
@@ -1,0 +1,1 @@
+export 'app_setting.dart';
